+++ Sources/parseConfigFile/displayServerFile.cpp
@@ -17,13 +17,13 @@ void	displayServerFile(std::vector<struct serverfile> conf)
 	int i;
 
 	i = 0;
-	while (i < conf.size())
+	while ((unsigned long)i < conf.size())
 	{
 		std::cout << "************************************************************" << std::endl;
 		std::cout << "Host : " << conf[i].host << std::endl;
 		{
 			int j = 0;
-			while (j < conf[i].listen.size())
+			while ((unsigned long)j < conf[i].listen.size())
 			{
 				std::cout << "listen : " << conf[i].listen[j] << std::endl;
 				j++;
@@ -31,7 +31,7 @@ void	displayServerFile(std::vector<struct serverfile> conf)
 		}
 		{
 			int j = 0;
-			while (j < conf[i].server_name.size())
+			while ((unsigned long)j < conf[i].server_name.size())
 			{
 				std::cout << "servername : " << conf[i].server_name[j] << std::endl;
 				j++;
@@ -39,7 +39,7 @@ void	displayServerFile(std::vector<struct serverfile> conf)
 		}
 		{
 			int j = 0;
-			while (j < conf[i].error_page.size())
+			while ((unsigned long)j < conf[i].error_page.size())
 			{
 				std::cout << "error_code: " << conf[i].error_page[j].first;
 				std::cout << ", error_page: " << conf[i].error_page[j].second;
@@ -52,7 +52,7 @@ void	displayServerFile(std::vector<struct serverfile> conf)
 		std::cout << "Index : " << conf[i].index << std::endl;
 		{
 			int j = 0;
-			while (j < conf[i].location.size())
+			while ((unsigned long)j < conf[i].location.size())
 			{
 				std::cout << "-------------------" << "\n" << "Location : " << std::endl;
 				std::cout << "pattern : " << conf[i].location[j].pattern << std::endl;
@@ -65,21 +65,21 @@ void	displayServerFile(std::vector<struct serverfile> conf)
 //				std::cout << "cgi_path : " << conf[i].location[j].cgi_path << std::endl;
 				std::cout << "redirect : " << conf[i].location[j].redirect << std::endl;
 				int k = 0;
-				while (k < conf[i].location[j].error_page.size())
+				while ((unsigned long)k < conf[i].location[j].error_page.size())
 				{
 					std::cout << "error code : " << conf[i].location[j].error_page[k].first << std::endl;
 					std::cout << "error file : " << conf[i].location[j].error_page[k].second << std::endl;
 					k++;
 				}
 				k = 0;
-				while (k < conf[i].location[j].cgi.size())
+				while ((unsigned long)k < conf[i].location[j].cgi.size())
 				{
 					std::cout << "cgi ext   : " << conf[i].location[j].cgi[k].first << std::endl;
 					std::cout << "cgi prog  : " << conf[i].location[j].cgi[k].second << std::endl;
 					k++;
 				}
 				k = 0;
-				while (k < conf[i].location[j].allow.size())
+				while ((unsigned long)k < conf[i].location[j].allow.size())
 				{
 					std::cout << "allow : " << conf[i].location[j].allow[k] << "\n";
 					k++;
